scraper/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.11.3
fastapi==0.115.12
feedparser==6.0.11
frozendict==2.4.6
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
httptools==0.6.4
idna==3.10
lxml==5.4.0
multidict==6.4.4
multitasking==0.0.11
numpy==2.3.0
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
prometheus_client==0.22.1
propcache==0.3.2
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.4
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.63
transformers>=4.0.0
torch>=1.13.0